fixed; undefined values are retained
We are now using $unset when a path is set to undefined.
This means the in the db your property will no longer be
set to null, it will be removed instead. This lets us
retain the intention of setting the val to undefined as
well as conserves db space.

closes #519