refactor field selection
There is now only one method for selecting/deselecting fields:

    query.select()

The `select` method accepts either an object or a string.

Passing an object allows for traditional MongoDB selection syntax.

    query.select({ included: 1, exlcuded: 0 })

Passing a string works as before but now includes the additional
ability to exclude fields by prefixing the key name with a "-".

    query.select('included -excluded')