added; document versioning support
When updating an embedded doc property the update object
specifies the path using positional notation:

    doc.comments[0].title = 'changed';
    doc.save() ->
      Model.update(
          { _id: doc._id }
        , { $set: { 'comments.0.title': 'changed' }} <--
      )

This is an unsafe operation in that another update may have
occurred between the time we queried the document and
saved. If the other update changed the underlying position of the
elements we refer to in our positional notation clause we will be
modifing the incorrect element. To mitigate the issue this commit
introduces document versioning.

Each document now receives a schema-configurable version key.
This value is incremented whenever a modification to an array
potentially changes any array's elements location.

When saving a document that requires use of positional notation
we now also send along the private document version counter in
the where clause set to the value received when first queried.