added; capped collection support
To enable a capped collection, set the `capped` option
in your schema to an object. The object has the following
properties:

`size`: the size in bytes of the created collection (required)
`max`:  optionally specify the max number of docs in the collection
`autoIndexId`: optionally create a unique index on _id (default false)

    new Schema({..}, { capped: { size: 100000 }});

You may optionally also set `capped` to a number which will be used
as the collection size option.

    new Schema({..}, { capped: 100000 });

@see http://www.mongodb.org/display/DOCS/Capped+Collections#CappedCollections-Options

closes #938