partial revert of eca2607
Instead of erroring when the model for an adhoc path cannot
be determined, use the current model. This opens the door
to populating path combinations of two+ separate schemas which
are now populated, e.g. populate('owner.avatar'), avatar being
a property of an owner field that was also previously populated.

relates to eca26073e3c6e1895ea96632669608b68d1f323c

closes #1359