model.findAndUpdateOne.test.js tests passing
all of these tests are now passing. A couple updates were made to the
tests themselves for this in order to make them consistent with the new
semantics in mquery. (namely, that _updateArgs is now _update and if you
pass no arguments to the update functions, _update is undefined, not an
empty object)