Improve null handling in Query.prototype._castUpdate
Query.prototype._castUpdate assumes that property values of obj are
non-null when checking the constructor of the value.  This will cause a
TypeError when the values are null, which can legitimately occur during
an update with the overwrite option.  Handle this case by checking if
the value is truthy before checking the name of the constructor.

Signed-off-by: Kevin Locke <klocke@quantpost.com>