Authentication will now still occur when auth method is MONGODB-X509 and no password is supplied.
MONGODB-X509 authentication uses certificates to authenticate clients instead of a password. This means that you can leave the password field blank. The problem is that currently if you leave the password as blank than authentication doesn't happen at all and you do not have permissions to execute anything on the database.

I have added a method that checks if the current authentication method needs a password and, if it does not, still calls the mongo driver's authenticate method.