Fixed a bug where username from options would not be accepted.
If the password was null than the username from the options object was ignored even when using an authentication method that does not need a password such as MONGODB-X509. This caused the authentication to fallback to the connection URL string for a username. If there was no username supplied in the connection URL string than the authentication would fail.