fix(populate): handle deeply nested virtual populate
The `subPopulate()` function was proving to be a weak abstraction, primarily because it created a strange ordering for initializing populated docs. For A -> B -> C, it would initialize B -> C -> A (as opposed to C -> B -> A) because it treated top-level populate differently from sub populates. This caused deeply nested virtual population to fail, because 'B' would not know about 'C' when getting initialized.

Fix #4278