Update subdocs.jade to conform to new highlighting
as mentioned in #6038 the formatting for code blocks is broken on 4 pages in the docs. I believe this brings the subdoc page inline with the working pages. If it works for you I'll do the remaining 3 files to match.

actually making the file match

I missed a couple details in my original commit. sorry for the extra notifications.

whitespace on line 10

extra whitespace removed. jade didn't complain about it, but it irked me.