docs(migration): certain in passwords may cause connection failures
In addition to domain sockets, certain passwords may trigger
connection failures in Mongoose 5. This can especially be a problem
if you use simpler passwords in testing and complex ones in production!

Links to related docs and a code sample to resolve the issue are
provided.

Related issue: https://github.com/Automattic/mongoose/issues/6044