update connections.html
I noticed the mongoose.connect() doesn't work properly without adding {useNewUrlParser: true} as an option. So I thought of updating this since many new users are facing unecessary trouble for this. This fix was actually recommended by @nax3t in the Q&A for the web developer bootcamp course in udemy. If this kind of PR is already rejected then please point me to the reason why if not then plzz merge it and also point me in the right direction if i can do anything more to improve this PR.
Thank you.