Update connections.pug
Make it clear that mongoose.connection.on('error') does not fire once connection is lost any time after the initial connection. The text seems to suggest, to me at least, that it's the code to use to listen to drops in connection after the initial connection.
Only 3 Lines were added: 148-150.