fix: TypeScript Intellisense
This change aims to fix intellisense for developing type definitions in
mongoose.

In `.ts` files, we currently get a compilation warning from TypeScript
when trying to `import mongoose from 'mongoose'`:

```
Cannot find module 'mongoose' or its corresponding type declarations.ts(2307)
```

As such, none of the intellisense in VS Code works correctly.

This change moves the TypeScript configuration out of our test file, and
into a `tsconfig.json`, where IDEs will be able to find it and use it to
enable Intellisense for better TypeScript typings development.