fix(connection): fix promise chaining for openUri
When a user calls openUri and attaches more than one callback to its
return value with 'then' method, it should be able to pass the result of
the first callback to the second one. Currently, openUri always pass
'undefined' to the second callback, making it behave differently from a
regular promise. This is unlikely to be an expected behaviour.

To fix it, modify openUri to propagate the return value of the
user-provided callback to the 'then' method of the promise.