fix: set default for dotted path projection
When a schema included a subdocument that had a default value of empty
object and that subdocument itself included some defaults, the
subdocument was always null if a query was run with a dotted path
projection to the subdocument's default values.

Check not only that the curent path being evaluated to have defaults
applied, but also check if the current path is part of the included
children in the projection.