Fix: simplify MergeType constraints
Fixes 16 errors in models.d.ts on Typescript 4.8 which is currently in
beta. Typescript 4.8 forbids assignment of an unconstrained type
parameter to one that disallows `null` or `undefined`:
https://github.com/microsoft/TypeScript/issues/49489

The alternative fix is to constrain *lots* of types in models.d.ts to
`Record<string | number, any>` -- that is, to forbid `null | undefined`,
but I'm not sure it's worth the effort.