prevValue tests: use Node 24 compatible assert
The prevValue tests used "nulls" in expected values. This comparison fails in Node 24, as the actual value is "undefined".

However, the original value seems to be "undefined" in previous Node versions too. Therefore it does not seem like a change in behaviour, but only a change in how strict the assert is. Changing the expected value to "undefined" works in both Node 24 and previous Node versions.